+++ anyblok_pyramid/tests/test_predicate.py
@@ -7,6 +7,7 @@
 # obtain one at http://mozilla.org/MPL/2.0/.
 from .testcase import PyramidDBTestCase
 from pyramid.response import Response
+from anyblok_pyramid import set_callable, get_db_name
 
 
 def uninstalled(request):
@@ -22,12 +23,22 @@ class TestViewPredicate(PyramidDBTestCase):
     def setUp(self):
         super(TestViewPredicate, self).setUp()
         self.installed_blok = 'anyblok-core'
+        self.need_anyblok_registry = True
+
+    def tearDown(self):
+        super(TestViewPredicate, self).tearDown()
+        set_callable(get_db_name)
 
     def add_route_and_views(self, config):
         config.add_route('test', '/test/')
         config.add_view(uninstalled, route_name='test')
         config.add_view(installed, route_name='test',
                         installed_blok=self.installed_blok)
+
+    def add_route_and_views2(self, config):
+        config.add_route('test', '/test/')
+        config.add_view(installed, route_name='test',
+                        need_anyblok_registry=self.need_anyblok_registry)
 
     def assetOK(self):
         resp = self.webserver.get('/test/', status=200)
@@ -58,16 +69,51 @@ class TestViewPredicate(PyramidDBTestCase):
         registry.upgrade(uninstall=['anyblok-io'])
         self.assetKO()
 
+    def test_need_anyblok_registry_ok(self):
+        self.includemes.append(self.add_route_and_views2)
+        webserver = self.init_web_server()
+        webserver.get('/test/', status=200)
+
+    def test_need_anyblok_registry_ko(self):
+        @set_callable
+        def get_db_name(request):
+            return 'wrong_db_name'
+
+        self.includemes.append(self.add_route_and_views2)
+        webserver = self.init_web_server()
+        webserver.get('/test/', status=404)
+
+    def test_need_anyblok_registry_ok2(self):
+        self.need_anyblok_registry = False
+
+        @set_callable
+        def get_db_name(request):
+            return 'wrong_db_name'
+
+        self.includemes.append(self.add_route_and_views2)
+        webserver = self.init_web_server()
+        webserver.get('/test/', status=200)
+
 
 class TestRoutePredicate(PyramidDBTestCase):
 
     def setUp(self):
         super(TestRoutePredicate, self).setUp()
         self.installed_blok = 'anyblok-core'
+        self.need_anyblok_registry = True
+
+    def tearDown(self):
+        super(TestRoutePredicate, self).tearDown()
+        set_callable(get_db_name)
 
     def add_route_and_views(self, config):
         config.add_route('test', '/test/', installed_blok=self.installed_blok)
         config.add_view(installed, route_name='test')
+
+    def add_route_and_views2(self, config):
+        config.add_route('test', '/test/')
+        config.add_view(installed, route_name='test',
+                        need_anyblok_registry=self.need_anyblok_registry)
 
     def assetOK(self):
         resp = self.webserver.get('/test/', status=200)
@@ -93,3 +139,28 @@ class TestRoutePredicate(PyramidDBTestCase):
         self.assetOK()
         registry.upgrade(uninstall=['anyblok-io'])
         self.webserver.get('/test/', status=404)
+
+    def test_need_anyblok_registry_ok(self):
+        self.includemes.append(self.add_route_and_views2)
+        webserver = self.init_web_server()
+        webserver.get('/test/', status=200)
+
+    def test_need_anyblok_registry_ko(self):
+        @set_callable
+        def get_db_name(request):
+            return 'wrong_db_name'
+
+        self.includemes.append(self.add_route_and_views2)
+        webserver = self.init_web_server()
+        webserver.get('/test/', status=404)
+
+    def test_need_anyblok_registry_ok2(self):
+        self.need_anyblok_registry = False
+
+        @set_callable
+        def get_db_name(request):
+            return 'wrong_db_name'
+
+        self.includemes.append(self.add_route_and_views2)
+        webserver = self.init_web_server()
+        webserver.get('/test/', status=200)